+++ scripts/analysis/common.py
@@ -1,0 +1,28 @@
+import pandas as pd
+
+
+def create_grouping_totals(
+    data: pd.DataFrame, group_column: str, exclude_cols: list[str]
+) -> pd.DataFrame:
+    """Create group totals as 'country'"""
+
+    dfs = []
+
+    for group in data[group_column].unique():
+        df_ = data.loc[lambda d: d[group_column] == group].copy()
+        df_["country"] = group
+        df_ = (
+            df_.groupby(
+                [c for c in df_.columns if c not in ["value"] + exclude_cols],
+                observed=True,
+                dropna=False,
+            )["value"]
+            .sum()
+            .reset_index()
+        )
+
+        dfs.append(df_)
+
+    groups = pd.concat(dfs, ignore_index=True)
+
+    return pd.concat([data, groups], ignore_index=True)
